bugfix: endType字段值异常问题修复
原因
endType字段为short类型，实际值为事件码，超出short范围

修改
1. 将所有short修改为int
2. 增加错误日志

测试
联调OK

+++ service-core/src/main/java/com/juphoon/rtc/datacenter/entity/po/acdstat/AcdAgentOpStatPO.java
@@ -38,22 +38,22 @@ public class AcdAgentOpStatPO extends AcdCommonPO {
     /**
      * 事件类型
      */
-    private Short eventType;
+    private Integer eventType;
 
     /**
      * 事件编号
      */
-    private Short eventNum;
+    private Integer eventNum;
 
     /**
      * 事件结束类型
      */
-    private Short endType;
+    private Integer endType;
 
     /**
      * 扩展状态(示忙子状态小休等)
      */
-    private Short extStatus;
+    private Integer extStatus;
 
     /**
      * 用event初始化部分字段
